"Agrego variables a Paciente, y clases Prioridad, TipoDeSintoma y Sintoma"

+++ test/unit/triage/PersonaControllerSpec.groovy
@@ -6,14 +6,21 @@ import grails.test.mixin.*
 import spock.lang.*
 
 @TestFor(PersonaController)
-@Mock(Persona)
+@Mock([Persona,Paciente])
 class PersonaControllerSpec extends Specification {
 
     def populateValidParams(params) {
         assert params != null
-        // TODO: Populate valid properties like...
-        //params["name"] = 'someValidName'
+        params["nombre"] = 'unNombre'
+		params["apellido"] = 'unApellido'
+		params["fechaDeNacimiento"] = '21/03/1987'
     }
+	
+	def populateValidJSONParams() {
+		request.JSON.nombre = 'unNombre'
+		request.JSON.apellido = 'unApellido'
+		request.JSON.fechaDeNacimiento = '21/03/1987'
+	}
 
     void "Test the index action returns the correct model"() {
 
@@ -44,7 +51,7 @@ class PersonaControllerSpec extends Specification {
             model.personaInstance!= null
             view == 'create'
 
-        /*when:"The save action is executed with a valid instance"
+        when:"The save action is executed with a valid instance"
             response.reset()
             populateValidParams(params)
             persona = new Persona(params)
@@ -54,7 +61,7 @@ class PersonaControllerSpec extends Specification {
         then:"A redirect is issued to the show action"
             response.redirectedUrl == '/persona/show/1'
             controller.flash.message != null
-            Persona.count() == 1*/
+            Persona.count() == 1
     }
 
     void "Test that the show action returns the correct model"() {
@@ -108,7 +115,7 @@ class PersonaControllerSpec extends Specification {
             view == 'edit'
             model.personaInstance == persona
 
-        /*when:"A valid domain instance is passed to the update action"
+        when:"A valid domain instance is passed to the update action"
             response.reset()
             populateValidParams(params)
             persona = new Persona(params).save(flush: true)
@@ -116,7 +123,7 @@ class PersonaControllerSpec extends Specification {
 
         then:"A redirect is issues to the show action"
             response.redirectedUrl == "/persona/show/$persona.id"
-            flash.message != null*/
+            flash.message != null
     }
 
     void "Test that the delete action deletes an instance if it exists"() {
@@ -127,20 +134,30 @@ class PersonaControllerSpec extends Specification {
             response.redirectedUrl == '/persona/index'
             flash.message != null
 
-        /*when:"A domain instance is created"
+        when:"A domain instance is created"
             response.reset()
             populateValidParams(params)
             def persona = new Persona(params).save(flush: true)
 
         then:"It exists"
-            Persona.count() == 1*/
+            Persona.count() == 1
 
-        /*when:"The domain instance is passed to the delete action"
+        when:"The domain instance is passed to the delete action"
             controller.delete(persona)
 
         then:"The instance is deleted"
             Persona.count() == 0
             response.redirectedUrl == '/persona/index'
-            flash.message != null*/
+            flash.message != null
     }
+	
+	void "Testeo que al crear una persona tambien se crea un paciente relacionado"(){
+		when:"Creo una persona y la persisto"
+			populateValidJSONParams()
+			controller.ajaxSave()
+
+		then:"Tambien se crea un paciente asociado y se persiste"
+			Persona.count() == 1
+			Paciente.count() == 1
+	}
 }
